CONSOLE-5196: Fix Playwright e2e flakes across multiple test suites
Address several sources of flaky Playwright e2e test failures:

1. Alertmanager parallel worker interference: Merge receivers.spec.ts
   into alertmanager.spec.ts so all tests that mutate the shared
   alertmanager-main secret run on a single worker.

2. warmupSPA resilience: Wrap the goto + heading check in toPass()
   with retry intervals so a single slow cluster response doesn't
   cause immediate failure.

3. a11y test stabilization: Wait for loading indicators to clear
   before running axe-core analysis. Upgrade toBeAttached() checks
   to toBeVisible() in assertLoaded callbacks so pages are fully
   rendered before a11y checks run.

4. Topology test timeouts: Add test.setTimeout(300_000) to tests
   that create workloads via git URL, since validation + deployment
   regularly exceeds the 120s default.

5. Topology application name field: Handle the conditional rendering
   of the application name input by selecting "Create application"
   from the dropdown when existing applications are present.

6. Perspective-switcher cleanup: Add afterAll to remove the Console
   operator perspectives patch, preventing cluster-wide side effects
   from leaking into other tests.
