+++ frontend/e2e/pages/base-page.ts
@@ -23,8 +23,10 @@ export async function setEditorContent(page: Page, content: string): Promise<voi
 }
 
 export async function warmupSPA(page: Page): Promise<void> {
-  await page.goto('/', { waitUntil: 'domcontentloaded', timeout: 60_000 });
-  await expect(page.getByTestId('page-heading')).toBeVisible({ timeout: 30_000 });
+  await expect(async () => {
+    await page.goto('/', { waitUntil: 'domcontentloaded', timeout: 60_000 });
+    await expect(page.getByTestId('page-heading')).toBeVisible({ timeout: 30_000 });
+  }).toPass({ intervals: [1_000, 2_000, 5_000], timeout: 90_000 });
 }
 
 export default abstract class BasePage {
